Reorganize controllers code to separate eventHandlers into file
