+++ controllers/workspace/devworkspace_controller.go
@@ -700,69 +700,6 @@ func (r *DevWorkspaceReconciler) getWorkspaceId(ctx context.Context, workspace *
 	}
 }
 
-// Mapping the pod to the devworkspace
-func dwRelatedPodsHandler(obj client.Object) []reconcile.Request {
-	labels := obj.GetLabels()
-	if _, ok := labels[constants.DevWorkspaceNameLabel]; !ok {
-		return []reconcile.Request{}
-	}
-
-	//If the dewworkspace label does not exist, do no reconcile
-	if _, ok := labels[constants.DevWorkspaceIDLabel]; !ok {
-		return []reconcile.Request{}
-	}
-
-	return []reconcile.Request{
-		{
-			NamespacedName: types.NamespacedName{
-				Name:      labels[constants.DevWorkspaceNameLabel],
-				Namespace: obj.GetNamespace(),
-			},
-		},
-	}
-}
-
-func (r *DevWorkspaceReconciler) dwPVCHandler(obj client.Object) []reconcile.Request {
-	// Check if PVC is owned by a DevWorkspace (per-workspace storage case)
-	for _, ownerref := range obj.GetOwnerReferences() {
-		if ownerref.Kind != "DevWorkspace" {
-			continue
-		}
-		return []reconcile.Request{
-			{
-				NamespacedName: types.NamespacedName{
-					Name:      ownerref.Name,
-					Namespace: obj.GetNamespace(),
-				},
-			},
-		}
-	}
-
-	// TODO: Label PVCs used for workspace storage so that they can be cleaned up if non-default name is used.
-	// Otherwise, check if common PVC is deleted to make sure all DevWorkspaces see it happen
-	if obj.GetName() != wkspConfig.GetGlobalConfig().Workspace.PVCName || obj.GetDeletionTimestamp() == nil {
-		// We're looking for a deleted common PVC
-		return []reconcile.Request{}
-	}
-	dwList := &dw.DevWorkspaceList{}
-	if err := r.Client.List(context.Background(), dwList, &client.ListOptions{Namespace: obj.GetNamespace()}); err != nil {
-		return []reconcile.Request{}
-	}
-	var reconciles []reconcile.Request
-	for _, workspace := range dwList.Items {
-		storageType := workspace.Spec.Template.Attributes.GetString(constants.DevWorkspaceStorageTypeAttribute, nil)
-		if storageType == constants.CommonStorageClassType || storageType == constants.PerUserStorageClassType || storageType == "" {
-			reconciles = append(reconciles, reconcile.Request{
-				NamespacedName: types.NamespacedName{
-					Name:      workspace.GetName(),
-					Namespace: workspace.GetNamespace(),
-				},
-			})
-		}
-	}
-	return reconciles
-}
-
 func (r *DevWorkspaceReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	setupHttpClients()
 
@@ -794,7 +731,7 @@ func (r *DevWorkspaceReconciler) SetupWithManager(mgr ctrl.Manager) error {
 		Watches(&source.Kind{Type: &corev1.Pod{}}, handler.EnqueueRequestsFromMapFunc(dwRelatedPodsHandler)).
 		Watches(&source.Kind{Type: &corev1.PersistentVolumeClaim{}}, handler.EnqueueRequestsFromMapFunc(r.dwPVCHandler)).
 		Watches(&source.Kind{Type: &controllerv1alpha1.DevWorkspaceOperatorConfig{}}, handler.EnqueueRequestsFromMapFunc(emptyMapper), configWatcher).
-		WithEventFilter(predicates).
+		WithEventFilter(devworkspacePredicates).
 		WithEventFilter(podPredicates).
 		Complete(r)
 }
